refactors to organize `data` crate
